+++ lib/src/types/index.ts
@@ -1,3 +1,4 @@
 export { default as Constructor } from './constructor';
 export { default as Provider } from './provider';
+export { default as Serializable } from './serializable';
 export { default as Util } from './util';