+++ lib/src/types/serializable.ts
@@ -1,8 +1,12 @@
-import { Constructor } from 'serialazy';
+import { Constructor } from './constructor';
 
 /** Generic serializable */
 export default abstract class Serializable {
 
+    /**
+     * Create new instance
+     * @param fields Instance fields
+     */
     public static create<T extends Serializable>(
         this: Constructor<T>,
         fields?: T