+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Add `Serializable`, an abstract base class for serializables
 * Async serialization/deserialization (`deflate.resolve` and `inflate.resolve`)
 * Serialization/deserialization to/from a POJO (`as` and `toPojo` options)
 * Support projections (`projection` and `fallbackToDefaultProjection` options)